Refactor: Replace MeasureUnit with custom TypeUnit enum
This commit replaces the `android.icu.util.MeasureUnit` class with a custom enum called `TypeUnit`. This change was made to support pressure units that are not available in the `MeasureUnit` class.

Key changes:
- Introduced `TypeUnit.kt` which includes pressure units like Bar, Pascal, and Atmosphere.
- Refactored the entire app to use `TypeUnit` instead of `MeasureUnit`.
- Updated navigation arguments to pass unit names (String) instead of indices (Int) for better readability and stability.
- Added a new "Pressure" conversion category.
- Included the corresponding icon and string resources for the new category.
- Renamed `Database.kt` to `AppDatabase.kt` for clarity.

+++ app/src/main/java/com/h3r3t1c/wearunitconverter/ui/compose/nav/Nav.kt
@@ -29,8 +29,8 @@ fun Nav(destination: String) {
             composable(NavDestination.CONVERT_PATH) {
                 val type = it.arguments?.getString("type")
                 val number = it.arguments?.getString("inputNumber")
-                val firstUnit = it.arguments?.getString("from")!!.toInt()
-                val secondUnit = it.arguments?.getString("to")!!.toInt()
+                val firstUnit = it.arguments?.getString("from")!!
+                val secondUnit = it.arguments?.getString("to")!!
                 ConvertMainScreen(navController, type!!, number!!, firstUnit, secondUnit)
             }
             composable(NavDestination.CONVERT_TO_OPTIONS_PATH) {
